dublin_bus/Dublin_bus/pip_dublinbus_requirements_20190618.txt
certifi==2019.3.9
Django==2.0
emoji==0.5.1
mysql-connector==2.2.9
numpy==1.16.4
pandas==0.24.2
pip==19.1.1
PyMySQL==0.9.3
python-dateutil==2.8.0
pytz==2019.1
setuptools==41.0.1
six==1.12.0
sqlparse==0.3.0
wheel==0.33.4
wincertstore==0.2
